backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
pykrx==1.0.45
requests==2.32.3
python-multipart==0.0.9
google-generativeai==0.7.2
apscheduler==3.10.4
